shadysim/requirements.txt
pyscard
pyserial
pytlv
cmd2>=1.5
jsonpath-ng
construct>=2.10.70
bidict
pyosmocom>=0.0.8
pyyaml>=5.1
termcolor
colorlog
pycryptodomex
cryptography
git+https://github.com/osmocom/asn1tools
packaging
git+https://github.com/hologram-io/smpp.pdu
smpp.twisted3 @ git+https://github.com/jookies/smpp.twisted
pycryptodome